refactor: 위시리스트의 위시 조회 API 위치 변경
# Conflicts:
#	backend/src/main/java/com/pickeat/backend/wish/ui/api/WishListApiSpec.java
