+++ backend/src/main/java/com/pickeat/backend/wish/ui/WishController.java
@@ -1,15 +1,18 @@
 package com.pickeat.backend.wish.ui;
 
 import com.pickeat.backend.global.auth.LoginUserId;
+import com.pickeat.backend.global.auth.ParticipantId;
 import com.pickeat.backend.wish.application.WishService;
 import com.pickeat.backend.wish.application.dto.request.WishRequest;
 import com.pickeat.backend.wish.application.dto.response.WishResponse;
 import com.pickeat.backend.wish.ui.api.WishApiSpec;
 import jakarta.validation.Valid;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -40,5 +43,15 @@ public class WishController implements WishApiSpec {
     public ResponseEntity<Void> deleteWish(@PathVariable("wishId") Long wishId, @LoginUserId Long userId) {
         wishService.deleteWish(wishId, userId);
         return ResponseEntity.noContent().build();
+    }
+
+    @Override
+    @GetMapping("/wishLists/{wishListId}/wishes")
+    public ResponseEntity<List<WishResponse>> getWishesInWishList(
+            @PathVariable("wishListId") Long wishListId,
+            @ParticipantId Long participantId
+    ) {
+        List<WishResponse> wishes = wishService.getWishes(wishListId, participantId);
+        return ResponseEntity.ok(wishes);
     }
 }
