+++ backend/src/main/java/com/pickeat/backend/wish/ui/WishListController.java
@@ -5,7 +5,6 @@ import com.pickeat.backend.global.auth.ParticipantId;
 import com.pickeat.backend.wish.application.WishListService;
 import com.pickeat.backend.wish.application.dto.request.WishListRequest;
 import com.pickeat.backend.wish.application.dto.response.WishListResponse;
-import com.pickeat.backend.wish.application.dto.response.WishResponse;
 import com.pickeat.backend.wish.ui.api.WishListApiSpec;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -52,15 +51,5 @@ public class WishListController implements WishListApiSpec {
     public ResponseEntity<List<WishListResponse>> getPublicWishLists() {
         List<WishListResponse> wishLists = wishListService.getPublicWishLists();
         return ResponseEntity.ok(wishLists);
-    }
-
-    @Override
-    @GetMapping("/wishLists/{wishListId}/wishes")
-    public ResponseEntity<List<WishResponse>> getWishesInWishList(
-            @PathVariable("wishListId") Long wishListId,
-            @ParticipantId Long participantId
-    ) {
-        List<WishResponse> wishes = wishListService.getWishes(wishListId, participantId);
-        return ResponseEntity.ok(wishes);
     }
 }
